# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Tapture backend
  version: '1'
  description:
    Browsers may call the API cross-origin only from the exact origins in CORS_ORIGINS. Those origins
    receive Access-Control-Allow-Origin and Vary Origin, and preflights answer 204 allowing GET, POST, PATCH
    and DELETE with the Authorization, Content-Type, X-Api-Version and Idempotency-Key headers. Other
    preflights answer 403. Credentials are never allowed. An empty CORS_ORIGINS sends no CORS headers.
paths:
...
  /api/v1/auth/register:
    post:
      description:
        Processes registration without revealing whether an address exists. Existing and new addresses
        receive the same accepted response and no account identifier.
      responses:
        '200':
          description: Accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accepted'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: *a1
                password:
                  type: string
                  minLength: 12
                organisationId: *a1
                invitationToken: *a1
              required:
                - email
                - password
                - organisationId
              additionalProperties: false
      parameters: []
  /api/v1/auth/change-password:
    post:
      responses:
        '204':
          description: Changed.
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword: *a1
                nextPassword:
                  type: string
                  minLength: 12
              required:
                - currentPassword
                - nextPassword
              additionalProperties: false
      parameters: []
  /api/v1/auth/reset:
    post:
      responses:
        '200':
          description: Accepted. The body does not reveal whether the address exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accepted'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  properties:
                    email: *a1
                    organisationId: *a1
                  required:
                    - email
                    - organisationId
                  additionalProperties: false
                - type: object
                  properties:
                    token: *a1
                    password:
                      type: string
                      minLength: 12
                  required:
                    - token
                    - password
                  additionalProperties: false
      parameters: []
  /api/v1/auth/login:
    post:
      description:
        Signs into this organisation and enrols a device. organisationId is optional when the deployment
        hosts exactly one organisation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: *a1
                password: *a1
                deviceId: *a1
                organisationId: *a1
              required:
                - email
                - password
                - deviceId
              additionalProperties: false
      responses:
        '200':
          description: Token pair.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters: []
  /api/v1/auth/refresh:
    post:
      responses:
        '200':
          description: Rotated pair.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: *a1
              required:
                - refreshToken
              additionalProperties: false
      parameters: []
  /api/v1/auth/logout:
    post:
      responses:
        '204':
          description: Signed out. Repeating it succeeds.
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: *a1
              required:
                - refreshToken
              additionalProperties: false
      parameters: []
...
  /api/v1/org/users:
    get:
      responses:
        '200':
          description: Page of users.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters:
        - name: cursor
          in: query
          schema: *a1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
    post:
      responses:
        '201':
          description: Invited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: *a1
                role: &a2
                  type: string
                  enum:
                    - administrator
                    - project_manager
                    - reviewer
                    - field_operator
              required:
                - email
                - role
              additionalProperties: false
      parameters: []
  /api/v1/org/users/{id}:
    patch:
      responses:
        '204':
          description: Role or status changed.
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: *a2
                status: &a5
                  type: string
                  enum:
                    - active
                    - disabled
                    - invited
              required: []
              additionalProperties: false
      parameters:
        - name: id
          in: path
          required: true
          schema: *a1
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - unauthorized
                - invalid_credentials
                - forbidden
                - not_found
                - conflict
                - payload_too_large
                - rate_limited
                - quota_exceeded
                - internal
                - unavailable
            message: *a1
            details:
              type: object
              additionalProperties: true
          required:
            - code
            - message
          additionalProperties: false
      required:
        - error
      additionalProperties: false
    Accepted:
      type: object
      properties:
        accepted:
          type: boolean
          enum:
            - true
      required:
        - accepted
      additionalProperties: false
    TokenPair:
      type: object
      properties:
        accessToken: *a1
        refreshToken: *a1
        expiresIn: *a4
      required:
        - accessToken
        - refreshToken
        - expiresIn
      additionalProperties: false
    User:
      type: object
      properties:
        id: *a1
        organisationId: *a1
        email: *a1
        role: *a2
        status: *a5
      required:
        - id
        - organisationId
        - email
        - role
        - status
      additionalProperties: false
    UserPage:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        nextCursor: *a6
      required:
        - users
        - nextCursor
      additionalProperties: false
    Invitation:
      type: object
      properties:
        userId: *a1
        invitationToken: *a1
      required:
        - userId
        - invitationToken
      additionalProperties: false
...
  responses:
    Error:
      description: Typed safe error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'